cpukit/libfs: Correct comparisons of different signedness
The GCC warning -Wsign-compare flagged these instances of comparing
signed and unsigned types. A common pair was size_t and int.

+++ cpukit/libfs/src/rfs/rtems-rfs-trace.c
@@ -117,7 +117,7 @@ rtems_rfs_trace_shell_command (int argc, char *argv[])
   rtems_rfs_trace_mask clear_value = 0;
   bool                 set = true;
   int                  arg;
-  int                  t;
+  size_t               t;
 
   for (arg = 1; arg < argc; arg++)
   {
